requirements.txt
biopython==1.85
pandas==2.3.3
obonet==1.1.1
